+++ src/Frontend/Dfe.Complete/Pages/Projects/TaskList/Tasks/DeclarationOfExpenditureCertificate/DeclarationOfExpenditureCertificate.cshtml.cs
@@ -1,0 +1,48 @@
+using Dfe.Complete.Domain.Enums;
+using MediatR;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
+
+namespace Dfe.Complete.Pages.Projects.TaskList.Tasks.DeclarationOfExpenditureCertificate
+{
+    public class DeclarationOfExpenditureCertificateModel(ISender sender, IAuthorizationService authorizationService, ILogger<DeclarationOfExpenditureCertificateModel> logger)
+    : BaseProjectTaskModel(sender, authorizationService, logger, NoteTaskIdentifier.DeclarationOfExpenditureCertificate)
+    {
+        [BindProperty(Name = "not-applicable")]
+        public bool? NotApplicable { get; set; }
+        [BindProperty(Name = "saved")]
+        public bool? Saved { get; set; }
+
+        [BindProperty(Name = "check-certificate")]
+        public bool? CheckCertificate { get; set; }
+
+        [BindProperty(Name = "date-received")]
+        public DateOnly? DateReceived { get; set; }
+
+        [BindProperty]
+        public Guid? TasksDataId { get; set; }
+        [BindProperty]
+        public ProjectType? Type { get; set; }
+        public override async Task<IActionResult> OnGetAsync()
+        {
+            await base.OnGetAsync();
+            Type = Project.Type;
+            TasksDataId = Project.TasksDataId?.Value;
+            if (Project.Type == ProjectType.Transfer)
+            {
+                NotApplicable = TransferTaskData.DeclarationOfExpenditureCertificateNotApplicable;
+                CheckCertificate = TransferTaskData.DeclarationOfExpenditureCertificateCorrect;
+                DateReceived = TransferTaskData.DeclarationOfExpenditureCertificateDateReceived;
+                Saved = TransferTaskData.DeclarationOfExpenditureCertificateSaved;
+            }
+            else
+            {
+                NotApplicable = ConversionTaskData.ReceiveGrantPaymentCertificateNotApplicable;
+                CheckCertificate = ConversionTaskData.ReceiveGrantPaymentCertificateCheckCertificate;
+                DateReceived = ConversionTaskData.ReceiveGrantPaymentCertificateDateReceived;
+                Saved = ConversionTaskData.ReceiveGrantPaymentCertificateSaveCertificate;
+            }
+            return Page();
+        }
+    }
+}